+++ hydra-cardano-api/src/Hydra/Cardano/Api/TxId.hs
@@ -1,3 +1,4 @@
+{-# OPTIONS_GHC -Wno-orphans #-}
 module Hydra.Cardano.Api.TxId where
 
 import Hydra.Cardano.Api.Prelude
@@ -5,6 +6,19 @@ import Hydra.Cardano.Api.Prelude
 import qualified Cardano.Crypto.Hash.Class as CC
 import qualified Cardano.Ledger.SafeHash as Ledger
 import qualified Cardano.Ledger.TxIn as Ledger
+import Cardano.Binary (ToCBOR(..), FromCBOR (..))
+
+-- missing CBOR instances
+
+instance ToCBOR TxId where
+  toCBOR = toCBOR . serialiseToRawBytes
+
+instance FromCBOR TxId where
+  fromCBOR = do
+    bs <- fromCBOR
+    case deserialiseFromRawBytes AsTxId bs of
+      Left err -> fail (show err)
+      Right v -> pure v
 
 -- * Type Conversions