+++ hydra-node/src/Hydra/Network/Message.hs
@@ -7,7 +7,7 @@ import Hydra.Prelude
 import Cardano.Binary (serialize')
 import Cardano.Crypto.Util (SignableRepresentation, getSignableRepresentation)
 import Hydra.Crypto (Signature)
-import Hydra.Ledger (IsTx, UTxOType)
+import Hydra.Ledger (IsTx (TxIdType), UTxOType)
 import Hydra.Network (NodeId)
 import Hydra.Snapshot (Snapshot, SnapshotNumber)
 
@@ -19,7 +19,7 @@ data Connectivity
 
 data Message tx
   = ReqTx {transaction :: tx}
-  | ReqSn {snapshotNumber :: SnapshotNumber, transactions :: [tx]}
+  | ReqSn {snapshotNumber :: SnapshotNumber, transactionIds :: [TxIdType tx]}
   | -- NOTE: We remove the party from the ackSn but, it would make sense to put it
     -- back as the signed snapshot is tied to the party and we should not
     -- consider which party sent this message to validate this snapshot signature.
@@ -29,19 +29,23 @@ data Message tx
     -- good idea to introduce the party in AckSn again or, maybe better, only
     -- the verification key of the party.
     AckSn {signed :: Signature (Snapshot tx), snapshotNumber :: SnapshotNumber}
-  deriving stock (Generic, Eq, Show)
-  deriving anyclass (ToJSON, FromJSON)
+  deriving stock (Generic)
+
+deriving instance (IsTx tx) => Eq (Message tx)
+deriving instance (IsTx tx) => Show (Message tx)
+deriving instance (IsTx tx) => ToJSON (Message tx)
+deriving instance (IsTx tx) => FromJSON (Message tx)
 
 instance IsTx tx => Arbitrary (Message tx) where
   arbitrary = genericArbitrary
 
-instance (ToCBOR tx, ToCBOR (UTxOType tx)) => ToCBOR (Message tx) where
+instance (ToCBOR tx, ToCBOR (UTxOType tx), ToCBOR (TxIdType tx)) => ToCBOR (Message tx) where
   toCBOR = \case
     ReqTx tx -> toCBOR ("ReqTx" :: Text) <> toCBOR tx
     ReqSn sn txs -> toCBOR ("ReqSn" :: Text) <> toCBOR sn <> toCBOR txs
     AckSn sig sn -> toCBOR ("AckSn" :: Text) <> toCBOR sig <> toCBOR sn
 
-instance (FromCBOR tx, FromCBOR (UTxOType tx)) => FromCBOR (Message tx) where
+instance (FromCBOR tx, FromCBOR (UTxOType tx), FromCBOR (TxIdType tx)) => FromCBOR (Message tx) where
   fromCBOR =
     fromCBOR >>= \case
       ("ReqTx" :: Text) -> ReqTx <$> fromCBOR